backend/requirements.txt
annotated-types==0.7.0
celery==5.6.2
django-celery-beat==2.8.1
django-celery-results==2.6.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
Brotli==1.1.0
browserforge==1.2.1
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==5.2.10
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-js-asset==3.0.1
djangorestframework==3.16.1
drf-yasg==1.21.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
instaloader==4.14
jiter==0.8.2
mccabe==0.7.0
MouseInfo==0.1.3
numpy==2.2.1
openai==1.58.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
playwright==1.49.1
pluggy==1.5.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pyclick==0.0.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyee==12.0.0
pyflakes==3.2.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyVirtualDisplay==3.0
PyYAML==6.0.2
redis==7.1.0
requests==2.32.3
selenium==4.27.1
selenium-wire==5.1.0
setuptools>=75.0.0
undetected-chromedriver==3.5.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
text-unidecode==1.3
TikTokApi==6.5.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.2
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
xlib==0.21
